requirements.txt
requests
lxml
scrapy
pandas
mysql-connector-python
sqlalchemy
matplotlib
seaborn
scikit-learn
xgboost
torch
#torch torchvision --index-url https://download.pytorch.org/whl/cu126
python-dotenv